create function for recent sort
