+++ app/client/src/ee/selectors/appIDESelectors.ts
@@ -1,0 +1,1 @@
+export * from "ce/selectors/appIDESelectors";
